Module Tezos_protocol_demo_counter__Protocol.Receipt

type t

receipts are used to notify users of operation application. We use a a simple string but we could use a structured datatype...

val create : string -> t

val to_string : t -> string

val encoding : t Tezos_protocol_environment_demo_counter__Environment.Data_encoding.t