Module Tezos_raw_protocol_demo_counter.Receipt

type t

receipts are used to notify users of operation application. We use a a simple string but we could use a structured datatype...

val create : string -> t

val to_string : t -> string

val encoding : t Tezos_base__TzPervasives.Data_encoding.t