Module Tezos_raw_protocol_demo_counter.Proto_params

type t = {

init_a : Tezos_protocol_environment_demo_counter__Environment.Int32.t;

init_b : Tezos_protocol_environment_demo_counter__Environment.Int32.t;

}

Constants of the demo protocol, passed at protocol activation. They defines the initial values of a and b

val default : t

val from_json : Tezos_protocol_environment_demo_counter__Environment.Data_encoding.json -> t

Create a t from a json structure of the form { 'init_a': int32, 'init_b': int32 }. Fields are optional with default value 0.